Add dev-toolbar switcher for Verbana and Baseline design variants.
