+++ src/platform/kbn-ui/chrome-layout/index.ts
@@ -30,6 +30,7 @@ export { CommonGlobalAppStyles } from './src/layouts/global_app_styles';
 
 export {
   DESIGN_EXPLORATION_BODY_ATTR,
+  DESIGN_EXPLORATION_VARIANT_ATTR,
   DESIGN_EXPLORATION_GAP,
   DESIGN_EXPLORATION_RADIUS_CONTAINER,
   DESIGN_EXPLORATION_RADIUS_CONTROL,
